Engine running on Marmalade!
git-svn-id: svn://10.0.50.55/legends/trunk/code@319 67207eb2-9daa-044d-9a34-b8567af74898

+++ src/GameLegends/Game.h
@@ -93,7 +93,7 @@ public:
 	void SortObject();
 	void SortObject3D();
 
-	void AddOldRoom( const Vector3D& _Pos, const char* _File, const Real _Scalar = Real(8) );
+	void AddOldRoom( const Vector3D& _Pos, const char* _File, const Real _Scalar = Real(16) );
 	void AddOldRoomMesh( const Vector3D& _Pos, const char* _File, const Real _Scalar = Real(2) );
 
 	void LoadMap();
